flatten format service facade into index.ts
